Move day/night WebView theming support into the library
