+++ core/src/main/kotlin/dev/hotwire/core/turbo/views/TurboWebView.kt
@@ -2,12 +2,16 @@ package dev.hotwire.core.turbo.views
 
 import android.annotation.SuppressLint
 import android.content.Context
+import android.content.res.Configuration
+import android.os.Build
 import android.util.AttributeSet
 import android.view.View
 import android.webkit.WebView
 import android.widget.FrameLayout
 import android.widget.FrameLayout.LayoutParams.MATCH_PARENT
+import androidx.webkit.WebSettingsCompat
 import androidx.webkit.WebViewCompat
+import androidx.webkit.WebViewFeature
 import com.google.gson.GsonBuilder
 import dev.hotwire.core.config.Hotwire
 import dev.hotwire.core.turbo.util.contentFromAsset
@@ -33,6 +37,7 @@ open class TurboWebView @JvmOverloads constructor(context: Context, attrs: Attri
         settings.userAgentString = "${Hotwire.config.userAgent} ${settings.userAgentString}"
         settings.setSupportMultipleWindows(true)
         layoutParams = FrameLayout.LayoutParams(MATCH_PARENT, MATCH_PARENT)
+        initDayNightTheming()
     }
 
     /**
@@ -87,5 +92,29 @@ open class TurboWebView @JvmOverloads constructor(context: Context, attrs: Attri
 
     private fun encodeArguments(vararg args: Any): String? {
         return args.joinToString(",") { gson.toJson(it) }
+    }
+
+    @Suppress("DEPRECATION")
+    private fun initDayNightTheming() {
+        if (Build.VERSION.SDK_INT < Build.VERSION_CODES.TIRAMISU) {
+            if (WebViewFeature.isFeatureSupported(WebViewFeature.FORCE_DARK_STRATEGY)) {
+                WebSettingsCompat.setForceDarkStrategy(
+                    settings,
+                    WebSettingsCompat.DARK_STRATEGY_WEB_THEME_DARKENING_ONLY
+                )
+            }
+
+            if (WebViewFeature.isFeatureSupported(WebViewFeature.FORCE_DARK)) {
+                when (isNightModeEnabled(context)) {
+                    true -> WebSettingsCompat.setForceDark(settings, WebSettingsCompat.FORCE_DARK_ON)
+                    else -> WebSettingsCompat.setForceDark(settings, WebSettingsCompat.FORCE_DARK_AUTO)
+                }
+            }
+        }
+    }
+
+    private fun isNightModeEnabled(context: Context): Boolean {
+        val currentNightMode = context.resources.configuration.uiMode and Configuration.UI_MODE_NIGHT_MASK
+        return currentNightMode == Configuration.UI_MODE_NIGHT_YES
     }
 }
